sms-categorization-app/requirements.txt
azure-functions
anthropic
aiohttp
hubspot-api-client
pytest
pytest-asyncio
